image_experiments/requirements.txt
# Core project dependencies
accelerate>=0.27.0
datasets>=2.14.0
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.1.0
Pillow>=9.5.0
pyarrow>=12.0.0
torch>=2.1.0
tqdm>=4.66.0
transformers>=4.37.0

# Optional utilities commonly used with PickScore
fire>=0.4.0
hydra-core>=1.3.2
rich>=13.3.2
submitit>=1.4.5
wandb>=0.12.21
